fix(post): RELEVANCE 검색 GIN 인덱스 사용 + 커서 정밀도 보강
- similarity() > 0.03 → search_text % :keyword 로 교체. 함수 호출 형태는
  gin_trgm_ops 인덱스를 못 타서 풀스캔이 되던 문제. 임계값은
  SET LOCAL pg_trgm.similarity_threshold = 0.03 으로 트랜잭션 스코프 지정.
- score 컬럼을 numeric(10,8) 로 캐스트하고 lastScore 를 BigDecimal 로
  통일. real(float4) → double 왕복에서 정밀도가 어긋나 커서 동등 비교
  (= :lastScore) 가 빗나가던 잠재 버그를 차단.
- searchPostsByRelevance 에 메서드 단위 @Transactional 추가로 클래스
  레벨 readOnly 를 오버라이드. SET LOCAL 을 readOnly 트랜잭션에서
  실행할 때의 드라이버/Hibernate 호환 문제를 회피.

+++ src/main/java/com/therapyCommunity_Vol1/backend/post/controller/PostController.java
@@ -25,6 +25,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Set;
 import org.springframework.http.HttpStatus;
@@ -114,7 +115,7 @@ public class PostController {
             @RequestParam String keyword,
             @RequestParam(required = false) TherapyArea therapyArea,
             @RequestParam(required = false) PostType postType,
-            @RequestParam(required = false) Double lastScore,
+            @RequestParam(required = false) BigDecimal lastScore,
             @RequestParam(required = false) Long lastId,
             @RequestParam(defaultValue = "10") int size
     ) {
